Refactor firewall utilities to centralized utils package
- Move firewall installation, collection, and reordering functions to pkg/utils/firewall.go
- Consolidate firewall_reorder.go, firewall_sync.go, firewall_utils.go into single utils/firewall.go
- Create new pkg/runner/firewall.go for firewall-specific command handlers
- Inject runCmdWithOutput function via SetFirewallCommandExecutor for proper dependency injection
- Update all command handlers to use utils.InstallFirewall() and utils.CollectFirewallRules()
- Maintain backward compatibility with existing firewall operations
